requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

langchain-google-genai
google-generativeai

numpy
scikit-learn
streamlit
tiktoken
faiss-cpu